Adding explicit casts to fix CS0266 introduced by Roslyn 16.9

+++ MBBSEmu/CPU/CPUCore.cs
@@ -1126,7 +1126,7 @@ namespace MBBSEmu.CPU
         [MethodImpl(OpcodeCompilerOptimizations)]
         private void Op_Cwd()
         {
-            Registers.DX = Registers.AX.IsBitSet(15) ? 0xFFFF : 0x0000;
+            Registers.DX = (ushort)(Registers.AX.IsBitSet(15) ? 0xFFFF : 0x0000);
         }
